requirements.txt
apache-airflow
boto3
python-dotenv
pyyaml
psycopg2-binary